requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
blinker==1.7.0
cachetools==5.3.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==3.1.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
evaluate==0.4.3
faiss-cpu==1.9.0
fastapi==0.108.0
ffmpy==0.3.1
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
gradio==4.13.0
gradio_client==0.8.0
greenlet==3.1.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.0.51
langchain-nvidia-ai-endpoints==0.3.5
langchain-text-splitters==0.3.2
langsmith==0.1.142
llama-cloud==0.1.4
llama-index==0.11.22
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-huggingface==0.1.3
llama-index-embeddings-nvidia==0.2.5
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-nvidia==0.2.7
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.13
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.54.3
orjson==3.10.11
packaging==23.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
proto-plus==1.23.0
protobuf==4.25.1
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.13
pyparsing==3.2.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.0
rouge_score==0.1.2
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.32.0.post1
streamlit==1.29.0
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.25.0
validators==0.22.0
watchdog==6.0.0
websockets==11.0.3
wordcloud==1.9.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
